move metabox code to new architecture
